moved files around to allow dependencies to be fetched during build
