+++ test/extract_metar.go
